% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviations.R
\docType{data}
\name{abbreviations}
\alias{abbreviations}
\title{abbreviations}
\format{a 'searchable' vector with 2,634 elements; names coorespond to 
words; values correspond to abbreviations}
\usage{
abbreviations
}
\description{
abbreviations
}
\keyword{datasets}
