% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/conform.R
\name{conform}
\alias{conform}
\alias{conform.character}
\alias{conform.data.frame}
\alias{conform.data.table}
\title{Conform}
\usage{
conform(x, ...)

\method{conform}{character}(x, fun = getOption("conform", function(x)
  lettercase::make_names(str_snake_case(abbreviate(stringi::stri_trim_both(x),
  tokenizer = "[_ ]")))), ...)

\method{conform}{data.frame}(x, ...)

\method{conform}{data.table}(x, ...)
}
\arguments{
\item{x}{object. Object to conform usually a data.frame, data.table or
character vector}

\item{...}{objects passed to the next method(s).}

\item{fun}{function to perform the conformation. The default is to look
for the global conformation function: \code{getOptions('conform')}}
}
\value{
data frame that has been
conformed to DP standards

same type of object as \code{x} with its property set by conform
}
\description{
Conform an object to internal standards
}
\details{
\code{conform} provides a very abstract way to ensure that all names
follow a predefined convention.

The default conformation ... \cr
  trims whitespace,
  abbreviates words,
  enforces snake_case.
  and makes_names
}
\note{
DP naming standards are lowercase, underscores('_') for whitespace
(not periods('.'), leading dot is retained

This function is \strong{idempotent} -- multiple application of the
function do not change the results.#'
}
\examples{
data(iris)
  conform(iris)
}
\author{
Christopher Brown
}

