% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/exact.R
\name{exact}
\alias{exact}
\title{Exact}
\usage{
exact(pattern)
}
\arguments{
\item{pattern}{string specifying a search pattern. Pattern modifiers can
be used, such as \code{\link{exact}}}
}
\description{
ensure that pattern matches the entire pattern and not any substring
}
\details{
Simply wraps a patterns between \code{"^"} and \code{"$"} ensuring that
  the regular expression matches the entire pattern.
}
\note{
This is a poor substitute for being implemented within \code{stringr}. See
  \url{https://github.com/hadley/stringr/issues/55}
}
\references{
\url{https://github.com/hadley/stringr/issues/55}
}

