% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup}
\alias{lookup}
\title{Lookup and entry in an object with names}
\usage{
lookup(x, dict = getOption("dictionary", abbreviations), missing = NA)
}
\arguments{
\item{x}{string; one-element character vector}

\item{dict}{object that has a names attribute that is used to look up the 
associated values.}

\item{missing}{value to return if the value cannot be found, (DEFAULT: NA)

Looks up and returns the item(s) in \code{dict} matching the pattern 
given by \code{x}. If there is no matrching name found, the value  provided 
in the \code{missing} argument is used.   

It is not uncommon to set \code{missing} to}
}
\value{
one or more values looked up by name in the dictionary. There is no 
  guarantee that there will be only one item returned per-lookup
}
\description{
Lookup and entry in an object with names
}
\examples{
  dict <- 1:26 
  names(dict) <- letters
  dict <- searchable(dict) 
  dict <- ignore.case(dict)
  
  lookup( ignore.case('M'), dict )    # 13
  lookup( 'M', dict )                 # NA
  lookup( 'M', dict, 'M' )            # M
  
}
\seealso{
\code{\link[searchable]{extract}} which is used internally to match 
  names
}
