% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/abbreviations.R
\docType{data}
\name{abbreviations}
\alias{abbreviations}
\title{abbreviations}
\format{a named character vector with 2,634 elements; names coorespond to
words; values correspond to abbreviations}
\usage{
abbreviations
}
\description{
abbreviations
}
\keyword{datasets}

